Реструктуризация проекта. Добавлены пакеты.
