+++ tests/__init__.py
@@ -2,6 +2,5 @@ from .test_coupling_coefficient import *
 
 from .tests import *
 from .tests_mcf_compression import *
-from .tests_mcf_nn import *
 # from .tests_parallel import *
 from .tests_unit import *
